requirements.txt
python-docx
PyPDF2
openpyxl
extract-msg
Pillow
pytesseract
xlrd
pandas